Move sort_by helper to submodule transaction
