+++ framework/src/modules/chain/logic/account.js
@@ -34,7 +34,6 @@ const __private = {};
  * @memberof logic
  * @see Parent: {@link logic}
  * @requires lodash
- * @requires helpers/sort_by
  * @requires helpers/bignum
  * @requires logic/block_reward
  * @param {Storage} storage
