+++ framework/src/modules/chain/submodules/transactions.js
@@ -19,7 +19,6 @@ const async = require('async');
 const {
 	CACHE_KEYS_TRANSACTION_COUNT,
 } = require('../../../../../framework/src/components/cache');
-const sortBy = require('../helpers/sort_by.js').sortBy;
 const TransactionPool = require('../logic/transaction_pool.js');
 const transactionTypes = require('../helpers/transaction_types.js');
 const Transfer = require('../logic/transfer.js');
@@ -42,7 +41,6 @@ __private.assetTypes = {};
  * @see Parent: {@link modules}
  * @requires bluebird
  * @requires lodash
- * @requires helpers/sort_by
  * @requires helpers/transaction_types
  * @requires logic/transaction_pool
  * @requires logic/transfer
@@ -201,7 +199,7 @@ __private.getPooledTransactions = function(method, filters, cb) {
 	}
 
 	// Sort the results
-	const sortAttribute = sortBy(filters.sort, { quoteField: false });
+	const sortAttribute = self.sortBy(filters.sort, { quoteField: false });
 
 	if (
 		sortAttribute.sortField === 'fee' ||
@@ -235,6 +233,127 @@ __private.getPooledTransactions = function(method, filters, cb) {
 		transactions: toSend,
 		count: transactions.length,
 	});
+};
+
+/**
+ * Validates sort options, methods and fields.
+ *
+ * @param {string|Object} sort
+ * @param {Object} [options]
+ * @param {string} options.fieldPrefix
+ * @param {string} options.sortField
+ * @param {string} options.sortMethod - asc / desc
+ * @param {Array} options.sortFields
+ * @returns {Object} {error} | {sortField, sortMethod}
+ * @todo Add description for the params
+ */
+Transactions.prototype.sortBy = function(sort, options) {
+	options = typeof options === 'object' ? options : {};
+	options.sortField = options.sortField || null;
+	options.sortMethod = options.sortMethod || null;
+	options.sortFields = Array.isArray(options.sortFields)
+		? options.sortFields
+		: [];
+
+	if (typeof options.quoteField === 'undefined') {
+		options.quoteField = true;
+	} else {
+		options.quoteField = Boolean(options.quoteField);
+	}
+
+	let sortField;
+	let sortMethod;
+
+	if (typeof sort === 'string') {
+		const [field, order] = sort.split(':');
+		sortField = field.replace(/[^\w\s]/gi, '');
+		sortMethod = order === 'desc' ? 'DESC' : 'ASC';
+	} else if (typeof sort === 'object') {
+		const keys = Object.keys(sort);
+
+		if (keys.length === 0) {
+			return self.sortBy('');
+		}
+		if (keys.length === 1) {
+			return self.sortBy(
+				`${keys[0]}:${sort[keys[0]] === -1 ? 'desc' : 'asc'}`,
+				options
+			);
+		}
+		const sortFields = [];
+		const sortMethods = [];
+		keys.forEach(key => {
+			const sortResult = self.sortBy(
+				`${key}:${sort[key] === -1 ? 'desc' : 'asc'}`,
+				options
+			);
+			sortFields.push(sortResult.sortField);
+			sortMethods.push(sortResult.sortMethod);
+		});
+		return { sortField: sortFields, sortMethod: sortMethods };
+	}
+	/**
+	 * Description of the function.
+	 *
+	 * @private
+	 * @todo Add param-tag and descriptions
+	 * @todo Add @returns tag
+	 * @todo Add description for the function
+	 */
+	function prefixField(prefixSortedField) {
+		if (!prefixSortedField) {
+			return prefixSortedField;
+		}
+		if (typeof options.fieldPrefix === 'string') {
+			return options.fieldPrefix + prefixSortedField;
+		}
+		if (typeof options.fieldPrefix === 'function') {
+			return options.fieldPrefix(prefixSortedField);
+		}
+		return prefixSortedField;
+	}
+
+	/**
+	 * Description of the function.
+	 *
+	 * @private
+	 * @todo Add param-tag and descriptions
+	 * @todo Add @returns tag
+	 * @todo Add description for the function
+	 */
+	function quoteField(quoteSortedField) {
+		if (quoteSortedField && options.quoteField) {
+			return `"${sortField}"`;
+		}
+		return quoteSortedField;
+	}
+
+	const emptyWhiteList = options.sortFields.length === 0;
+
+	const inWhiteList =
+		options.sortFields.length >= 1 &&
+		options.sortFields.indexOf(sortField) > -1;
+
+	if (sortField) {
+		if (emptyWhiteList || inWhiteList) {
+			sortField = prefixField(sortField);
+		} else {
+			return {
+				error: 'Invalid sort field',
+			};
+		}
+	} else {
+		sortField = prefixField(options.sortField);
+	}
+
+	if (!sortMethod) {
+		sortMethod = options.sortMethod;
+	}
+
+	return {
+		sortField: quoteField(sortField) || '',
+		sortMethod: sortField ? sortMethod : '',
+	};
 };
 
 // Public methods
